requirements.txt
numpy
tensorflow
pandas
python-dotenv
scikit-learn